convert pokemon.js to typescript and split out PokemonName
