+++ package.json
@@ -33,7 +33,7 @@
     "@types/react-modal": "^3.16.0",
     "@types/react-redux": "^7.1.25",
     "babel-loader": "^9.1.2",
-    "classnames": "^2.2.6",
+    "classnames": "^2.3.2",
     "core-js": "^3.30.2",
     "css-loader": "^6.8.1",
     "html-webpack-plugin": "^5.5.1",
